rafactor: restructure the project with layered organization
