+++ models/pdfreader_model.py
@@ -1,0 +1,6 @@
+from pydantic import BaseModel
+
+
+class PdfReaderModel(BaseModel):
+    """Data model for a question request."""
+    question: str
